- Moved StringUtils to auxiliary from containers - Updated PointerMath::Offset to be templated and replaced all occurences of it in code - Moved magic number logic to ICompiler instead of doing it in every derived compiler

+++ src/engine/input/input_filter.cc
@@ -37,7 +37,7 @@ namespace snuffbox
       size_t idx = (num_events_ - 1) * kEventSize_;
 
       void* block = 
-        foundation::PointerMath::Offset(buffer_, static_cast<int64_t>(idx));
+        foundation::PointerMath::Offset(buffer_, static_cast<intptr_t>(idx));
 
       *e = reinterpret_cast<const InputEvent*>(block);
